renamed old routes & tab routing
