+++ src/routing/challenge/ChallengeTabs.tsx
@@ -12,30 +12,37 @@ import React, { FC, useCallback, useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 import { Outlet, resolvePath, useResolvedPath } from 'react-router-dom';
 import NavigationTab from '../../components/core/NavigationTab/NavigationTab';
-import { useConfig, useEcoverse } from '../../hooks';
+import { useConfig } from '../../hooks';
 import useRouteMatch from '../../hooks/routing/useRouteMatch';
 import { FEATURE_COLLABORATION_CANVASES, FEATURE_COMMUNICATIONS_DISCUSSIONS } from '../../models/constants';
-import { buildAdminEcoverseUrl } from '../../utils/urlBuilders';
+import { buildAdminChallengeUrl } from '../../utils/urlBuilders';
 
 const routes = {
   discussions: 'discussions',
   community: 'community',
   dashboard: 'dashboard',
-  challenges: 'challenges',
+  opportunities: 'opportunities',
   canvases: 'canvases',
   settings: 'settings',
   context: 'context',
 };
 
-type EcoverseRoutesType = keyof typeof routes;
+type ChallengeRoutesKey = keyof typeof routes;
 
-export interface EcoverseTabsProps {
+export interface ChallengeTabsProps {
   communityReadAccess: boolean;
-  challengesReadAccess: boolean;
+  viewerCanUpdate: boolean;
+  ecoverseNameId: string;
+  challengeNameId: string;
 }
 
 // todo unify in one tab config component
-const EcoverseTabsNew: FC<EcoverseTabsProps> = ({ communityReadAccess, challengesReadAccess }) => {
+const ChallengeTabs: FC<ChallengeTabsProps> = ({
+  communityReadAccess,
+  viewerCanUpdate,
+  ecoverseNameId,
+  challengeNameId,
+}) => {
   const { t } = useTranslation();
   const { isFeatureEnabled } = useConfig();
   const resolved = useResolvedPath('.');
@@ -43,12 +50,8 @@ const EcoverseTabsNew: FC<EcoverseTabsProps> = ({ communityReadAccess, challenge
     () => Object.values(routes).map(x => resolvePath(x, resolved.pathname)?.pathname),
     [routes, resolved, resolvePath]
   );
-
-  // todo provided it as an input
-  const { ecoverseNameId, permissions } = useEcoverse();
-
   const tabValue = useCallback(
-    (route: EcoverseRoutesType | string) => resolvePath(route, resolved.pathname)?.pathname,
+    (route: ChallengeRoutesKey | string) => resolvePath(route, resolved.pathname)?.pathname,
     [resolved]
   );
 
@@ -61,7 +64,7 @@ const EcoverseTabsNew: FC<EcoverseTabsProps> = ({ communityReadAccess, challenge
     <>
       <Tabs
         value={currentTab}
-        aria-label="Ecoverse tabs"
+        aria-label="Challenge tabs"
         variant="scrollable"
         scrollButtons={'auto'}
         allowScrollButtonsMobile
@@ -70,43 +73,48 @@ const EcoverseTabsNew: FC<EcoverseTabsProps> = ({ communityReadAccess, challenge
           icon={<DashboardOutlined />}
           label={t('common.dashboard')}
           value={tabValue('dashboard')}
-          to={'dashboard'}
+          to={routes['dashboard']}
         />
-        <NavigationTab icon={<TocOutlined />} label={t('common.context')} value={tabValue('context')} to={'context'} />
+        <NavigationTab
+          icon={<TocOutlined />}
+          label={t('common.context')}
+          value={tabValue('context')}
+          to={routes['context']}
+        />
         <NavigationTab
           disabled={!communityReadAccess}
           icon={<GroupOutlined />}
           label={t('common.community')}
           value={tabValue('community')}
-          to={'community'}
+          to={routes['community']}
         />
         <NavigationTab
-          disabled={!challengesReadAccess}
           icon={<ContentPasteOutlined />}
-          label={t('common.challenges')}
-          value={tabValue('challenges')}
-          to={'challenges'}
+          label={t('common.opportunities')}
+          value={tabValue('opportunities')}
+          to={routes['opportunities']}
         />
         <NavigationTab
           disabled={!communityReadAccess || !isFeatureEnabled(FEATURE_COMMUNICATIONS_DISCUSSIONS)}
           icon={<ForumOutlined />}
           label={t('common.discussions')}
           value={tabValue('discussions')}
-          to={'discussions'}
+          to={routes['discussions']}
         />
         <NavigationTab
           disabled={!communityReadAccess || !isFeatureEnabled(FEATURE_COLLABORATION_CANVASES)}
           icon={<WbIncandescentOutlined />}
           label={t('common.canvases')}
           value={tabValue('canvases')}
-          to={'canvases'}
+          to={routes['canvases']}
         />
-        {permissions.viewerCanUpdate && (
+        {viewerCanUpdate && (
           <NavigationTab
             icon={<SettingsOutlined />}
             label={t('common.settings')}
             value={tabValue('settings')}
-            to={buildAdminEcoverseUrl(ecoverseNameId)}
+            /* can be provided with the tab config */
+            to={buildAdminChallengeUrl(ecoverseNameId, challengeNameId)}
           />
         )}
       </Tabs>
@@ -116,4 +124,4 @@ const EcoverseTabsNew: FC<EcoverseTabsProps> = ({ communityReadAccess, challenge
   );
 };
 
-export default EcoverseTabsNew;
+export default ChallengeTabs;
