+++ src/routing/opportunity/OpportunityTabs.tsx
@@ -6,114 +6,112 @@ import {
   TocOutlined,
   WbIncandescentOutlined,
 } from '@mui/icons-material';
-import { Tabs } from '@mui/material';
-import React, { FC, useMemo } from 'react';
+import { Box, Tabs } from '@mui/material';
+import React, { FC, useCallback, useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
-import { useResolvedPath } from 'react-router-dom';
+import { Outlet, resolvePath, useResolvedPath } from 'react-router-dom';
 import NavigationTab from '../../components/core/NavigationTab/NavigationTab';
-import { OpportunityContainerEntities } from '../../containers';
-import { useConfig, useOpportunity } from '../../hooks';
+import { useConfig } from '../../hooks';
 import useRouteMatch from '../../hooks/routing/useRouteMatch';
 import { FEATURE_COLLABORATION_CANVASES } from '../../models/constants';
 import { buildAdminOpportunityUrl } from '../../utils/urlBuilders';
 
 const routes = {
-  discussions: '/community/discussions',
-  community: '/community',
-  dashboard: '/dashboard',
-  projects: '/projects',
-  canvases: '/canvases',
-  settings: '/settings',
-  context: '/context',
-  root: '/',
+  community: 'community',
+  dashboard: 'dashboard',
+  projects: 'projects',
+  canvases: 'canvases',
+  settings: 'settings',
+  context: 'context',
 };
 
-export type OpportunityRoutesType = typeof routes;
+type OpportunityRoutesType = keyof typeof routes;
 
 export interface OpportunityTabsProps {
-  children: (e: {
-    pathGetter: (key: keyof typeof routes) => string;
-    urlGetter: (key: keyof typeof routes) => string;
-    tabName: string;
-    tabNames: OpportunityRoutesType;
-  }) => React.ReactNode;
-  entities: OpportunityContainerEntities;
+  communityReadAccess: boolean;
+  viewerCanUpdate: boolean;
+  ecoverseNameId: string;
+  challengeNameId: string;
+  opportunityNameId: string;
 }
-const createGetter = function <T>(r: T, url: string) {
-  return (key: keyof T) => `${url}${r[key]}`;
-};
-
-const OpportunityTabs: FC<OpportunityTabsProps> = ({ entities, children }) => {
-  // TODO use NavigationTabs and refactor the routing similar to the UserSettingsRoute.
-  const { pathname: path } = useResolvedPath('./');
+// todo unify in one tab config component
+const OpportunityTabs: FC<OpportunityTabsProps> = ({
+  viewerCanUpdate,
+  communityReadAccess,
+  ecoverseNameId,
+  challengeNameId,
+  opportunityNameId,
+}) => {
   const { t } = useTranslation();
-  const match = useRouteMatch(Object.values(routes).map(x => `${path}${x}/*`)); // Match underlying routes because of the discussions sub route.
-  const { ecoverseNameId, challengeNameId, opportunityNameId, permissions } = useOpportunity();
-  const urlGetter = useMemo(() => createGetter(routes, ''), []);
-  const pathGetter = useMemo(() => createGetter(routes, path), [path]);
-  const { communityReadAccess } = entities.permissions;
-
-  const tabNames = (Object.keys(routes) as Array<keyof OpportunityRoutesType>).reduce<OpportunityRoutesType>(
-    (acc, curr) => {
-      acc[curr] = pathGetter(curr);
-      return acc;
-    },
-    {} as OpportunityRoutesType
-  );
   const { isFeatureEnabled } = useConfig();
+  const resolved = useResolvedPath('.');
+  const matchPatterns = useMemo(
+    () => Object.values(routes).map(x => resolvePath(x, resolved.pathname)?.pathname),
+    [routes, resolved, resolvePath]
+  );
+  const tabValue = useCallback(
+    (route: OpportunityRoutesType | string) => resolvePath(route, resolved.pathname)?.pathname,
+    [resolved]
+  );
+
+  const routeMatch = useRouteMatch(matchPatterns);
+  const currentTab = useMemo(() => {
+    return routeMatch?.pattern?.path ?? tabValue('dashboard');
+  }, [routeMatch, routes]);
 
   return (
     <>
-      <Tabs value={match?.pathnameBase} aria-label="opportunity tabs">
+      <Tabs
+        value={currentTab}
+        aria-label="Opportunity tabs"
+        variant="scrollable"
+        scrollButtons={'auto'}
+        allowScrollButtonsMobile
+      >
         <NavigationTab
           icon={<DashboardOutlined />}
           label={t('common.dashboard')}
-          value={pathGetter('dashboard')}
-          to={urlGetter('dashboard')}
+          value={tabValue('dashboard')}
+          to={routes['dashboard']}
         />
         <NavigationTab
           icon={<TocOutlined />}
           label={t('common.context')}
-          value={pathGetter('context')}
-          to={urlGetter('context')}
+          value={tabValue('context')}
+          to={routes['context']}
         />
         <NavigationTab
           disabled={!communityReadAccess}
           icon={<GroupOutlined />}
           label={t('common.community')}
-          value={pathGetter('community')}
-          to={urlGetter('community')}
+          value={tabValue('community')}
+          to={routes['community']}
         />
         <NavigationTab
           icon={<ContentPasteOutlined />}
           label={t('common.projects')}
-          value={pathGetter('projects')}
-          to={urlGetter('projects')}
+          value={tabValue('projects')}
+          to={routes['projects']}
         />
-        {/* <NavigationTab
-          disabled={!communityReadAccess || !isFeatureEnabled(FEATURE_COMMUNICATIONS_DISCUSSIONS)}
-          icon={<ForumOutlined />}
-          label={t('common.discussions')}
-          value={pathGetter('discussions')}
-          to={urlGetter('discussions')}
-        /> */}
         <NavigationTab
           disabled={!communityReadAccess || !isFeatureEnabled(FEATURE_COLLABORATION_CANVASES)}
           icon={<WbIncandescentOutlined />}
           label={t('common.canvases')}
-          value={pathGetter('canvases')}
-          to={urlGetter('canvases')}
+          value={tabValue('canvases')}
+          to={routes['canvases']}
         />
-        {permissions.viewerCanUpdate && (
+        {viewerCanUpdate && (
           <NavigationTab
             icon={<SettingsOutlined />}
             label={t('common.settings')}
-            value={pathGetter('settings')}
+            value={tabValue('settings')}
+            /* can be provided with the tab config */
             to={buildAdminOpportunityUrl(ecoverseNameId, challengeNameId, opportunityNameId)}
           />
         )}
       </Tabs>
-      {children({ pathGetter, urlGetter, tabName: match?.pathnameBase || 'dashboard', tabNames })}
+      <Box paddingTop={3} />
+      <Outlet />
     </>
   );
 };
